Added routes file to handle middeware
